Renamed edge to leaf
