+++ packages/core/leaf.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 
 // These are overridden when running in cluster
 // They default to running locally
-const edgeURL = process.env.SP_EDGE_URL || 'http://localhost:3010'
+const leafURL = process.env.SP_LEAF_URL || 'http://localhost:3010'
 
 // In order for intercept to work, headers need to
 // be propagated to upstream requests
@@ -16,18 +16,18 @@ const propagateHeaders = (headers) =>
         return obj
       }, {})
 
-// Calls the edge service and obtains headers
-export const edgeResult = async (inHeaders) => {
+// Calls the leaf service and obtains headers
+export const leafResult = async (inHeaders) => {
     try {
         const headers = propagateHeaders(inHeaders)
-        const url = `${edgeURL}/api`
+        const url = `${leafURL}/api`
         const result = await axios({
             url,
             headers
         })
-        return { edge: { url, ...result.data, 'propagated-headers': headers } }
+        return { leaf: { url, ...result.data, 'propagated-headers': headers } }
     }
     catch (error) {
-        return { edge: { error: error.message } }
+        return { leaf: { error: error.message } }
     }
 }
